+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+#### 🐞 Fixes
+
+- Fixed an issue where env files (`.env`) would not be considered as task inputs. We've also updated
+  `env_file` to be an implicit input.
+
 ## 0.20.1
 
 #### 🐞 Fixes